Change PREBUILD_STRIP_BIN to path to strip binary
Problem: The strip binary wasn't found when we looked in the expected
location: `/usr/arm-linux-gnueabihf/bin/arm-linux-gnueabihf-strip`. It's
unclear why this is happening, but we need to change the environment
variable `PREBUILD_STRIP_BIN` so that it points at a valid binary.

Solution: We found three binaries that seem to look like `strip`:

- /usr/arm-linux-gnueabihf/bin/strip
- /usr/bin/arm-linux-gnueabihf-strip
- /usr/bin/x86_64-linux-gnu-strip

The first two are the exact same binary, confirmed with `sha256sum`, and
since their version string contains "crosstool" I'm guessing that
they're the right binary. Since both are probably valid, I've chosen the
one that starts with `/usr/arm-linux-gnueabihf` because it means that we
have to change fewer lines in the Dockerfile.

I've tested to ensure that `$PREBUILD_STRIP_BIN --version` correctly
runs the binary and returns basic version information, which suggests
that this should Actually Work, For Real This Time.