Add Android ARMv8 support
Problem: When installing Leveldown on the Google Pixel XL I noticed that
it's complaining that it cannot locate symbol "_Unwind_Resume".

Solution: I don't know why this is happening, but I'm wondering whether
a prebuild would solve the problem. This commit adds Android ARMv8
support, which I *think* should work as a prebuild on my device. I'm a
bit out of my element and wouldn't be surprised if this is the wrong
move.

See: https://github.com/Level/leveldown/issues/705