Install nodejs package from correct source (#18)
In our dockcross-based images, the 'nodejs' package is available in
both default debian sources and nodesource. Debian's package does
not include npm.

Not sure what changed or when, but it's fixed by increasing the
priority of nodesource. This in turn revealed that we were
targeting the wrong debian version; should be bullseye instead of
buster.

Closes #17.